functioning adding a venta and the log for it

+++ servicioVentas/src/main/java/ar/edu/um/programacion2/servicioVentas/model/Venta.java
@@ -33,7 +33,7 @@ public class Venta implements Serializable{
 	private Long tarjeta_id;
 	
 	@Column(name = "monto")
-	private Long monto;
+	private Double monto;
 	
 	@Column(name = "fecha")
 	private Date fecha;
